Add common superclass for IN predicates with constant and non-constant values
